refactor: `findSalesFiles` function to handle errors and improve readability
